refactor and add search bar to navbar, slight mobile navbar redesign
