+++ src/styles/Navbar.css
@@ -10,6 +10,7 @@
     display: flex;
     margin: auto;
     justify-content: space-between;
+    gap: 20px;
 }
 
 .navbar-title-container {
@@ -18,6 +19,7 @@
     align-items: center;
     gap: 3px;
     cursor: pointer;
+    margin-right: 5px;
 }
 
 .navbar-title {
@@ -28,6 +30,10 @@
     max-height: 13px;
     width: auto;
     margin-top: 4px;
+}
+
+.navbar-items-container {
+    flex-shrink: 0;
 }
 
 .navbar-item {
@@ -45,7 +51,8 @@
     padding: 5px;
     font-size: 0.7em;
     /* background-color: var(--blue800); */
-    background-color: rgba(255, 255, 255, 0.1);
+    /* background-color: rgba(255, 255, 255, 0.1); */
+    background-color: #2ab383;
     font-weight: 600;
     text-transform: uppercase;
     text-align: center;
@@ -99,5 +106,15 @@
     .navbar-container {
         justify-content: center;
         flex-direction: column;
+        gap: 0;
+        padding: 0;
+        padding-top: 15px;
+    }
+    .navbar-items-container {
+        width: 100%;
+        display: flex;
+        background: rgba(0, 0, 0, 0.2);
+        justify-content: center;
+        margin-top: 5px;
     }
 }
